+++ docs/source/modules.rst
@@ -5,3 +5,8 @@ censusutil
 ==========
 ..  autoclass:: censusutil.CensusUtil
     :members:
+
+censusviz
+==========
+..  autoclass:: censusviz.CensusViz
+    :members:
